// portafolio/src/app/components/models/user.ts
export interface User {
    id_user:          number;
    image_user:       string;
    description_user: string;
    name_user:        string;
    lastName_user:    string;
    email:            string;
    last_profession:  string;
    trade_profession: string;
    user_password:    string;
    phone_number:     string;
}